Calculate quad key externally
